Phase 2: live event stream over API keys + events scope
Lets external tools subscribe to a map's live edits with an API key, the
same SSE feed the web client uses (design: external_api_feature.md §8).

Backend:
- services/mapStream.ts: extract the SSE setup (headers, subscribe, heartbeat,
  presence snapshot, close cleanup) into a shared streamMapEvents() so the
  cookie route and the new key route behave identically. Safe for headless key
  clients — presence is only read, and close-time cleanup is session-guarded.
- maps.ts: GET /:mapId/events now delegates to the helper (unused subscribe/
  presence imports dropped).
- apiV1.ts: GET /api/v1/maps/:mapId/events — gated on the 'events' scope (a
  'read' key gets 403; a session passes through), then getMapAccess + stream.
  Scope hierarchy: an 'events' key can also hit the read endpoints.

Frontend:
- ApiKeysModal: access selector (Read only / Read + live events) on create,
  sent as scope; a scope badge per key in the list. New i18n + CSS.

README: document the live event stream endpoint, the 'events' scope, and the
published event taxonomy as the public contract.

Smoke-tested: read key -> events 403; no auth -> 401; read key still reads 200;
events key reads 200 and opens the stream (connect + presence.snapshot frames).
